// src/server/model/pLedgerMaster.ts
/**
 * 
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface PLedgerMaster { 
    id?: number;
    name?: string;
    printName?: string;
    ledgerGroupId?: number;
    ledgerGroupName?: string;
    openingBal?: number;
    gstin?: string;
    stateName?: string;
    stateId?: number;
    mobile?: string;
    email?: string;
    city?: string;
    disabled?: boolean;
    aliasName?: string;
    gstStatus?: string;
    systemGenerated?: boolean;
}